requirements.txt
pandas
streamlit
folium
streamlit_folium
plotly.express
